Refactor: changes switch to Formatted values
