+++ packages/grafana-data/src/field/index.ts
@@ -5,6 +5,6 @@ export * from './standardFieldConfigEditorRegistry';
 export * from './overrides/processors';
 export { FieldConfigOptionsRegistry } from './FieldConfigOptionsRegistry';
 
-export { applyFieldOverrides, validateFieldConfig } from './fieldOverrides';
+export { applyFieldOverrides, validateFieldConfig, applyRawFieldOverrides } from './fieldOverrides';
 export { getFieldDisplayValuesProxy } from './getFieldDisplayValuesProxy';
 export { getFieldDisplayName, getFrameDisplayName } from './fieldState';
